+++ src/main/java/com/sicca/repository/CultivoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CultivoRepository extends JpaRepository<CultivoEntity, Long> {
+public interface CultivoRepository extends JpaRepository<CultivoEntity, Integer> {
 }
