+++ src/main/java/com/sicca/repository/EspecieRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EspecieRepository extends JpaRepository<EspecieEntity, Long> {
+public interface EspecieRepository extends JpaRepository<EspecieEntity, Integer> {
 }
