+++ src/main/java/com/sicca/repository/ImagenRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImagenRepository extends JpaRepository<ImagenEntity, Long> {
+public interface ImagenRepository extends JpaRepository<ImagenEntity, Integer> {
 }
